Cleanup and implement unit testing.
As the index.js was getting more and more crowded I decided to move my function to a new folder 'utilities and require them as they needed.
This cleanup also helps the unit testing process.
As I am attempting to generate a large number of data points it would be
better to check if all units work as expected.
unit-testing using Mocha and Chai.